requirements.txt
beautifulsoup4
lxml
psycopg2
python-dateutil
python-telegram-bot
pytz
requests

SQLAlchemy
alembic